Extract symbol loading code from OrbitApp into new SymbolLoader class
The main goal is to reduce the size of `OrbitApp`. `OrbitApp` now uses the new
`SymbolLoader`, which calls into `OrbitApp` through the
`SymbolLoader::AppInterface` interface.
A couple of extra methods need to be added to `OrbitApp` for the communication
between `SymbolLoader` and `OrbitMainWindow`.

Test: Auto symbol loading on Trata and locally. Also try manual loading for
errors (rather, "Partial" cases) and try stopping the download of a module.
