+++ src/OrbitQt/orbitmainwindow.cpp
@@ -2032,11 +2032,8 @@ orbit_base::CanceledOr<void> OrbitMainWindow::DisplayStopDownloadDialog(
     case Result::kStopAndDisable: {
       metrics_uploader_->SendLogEvent(
           orbit_metrics_uploader::OrbitLogEvent::ORBIT_SYMBOL_DOWNLOAD_DISABLED_BY_USER);
-      orbit_client_symbols::QSettingsBasedStorageManager storage_manager;
-      absl::flat_hash_set<std::string> disabled_modules = storage_manager.LoadDisabledModulePaths();
-      disabled_modules.emplace(module->file_path());
-      storage_manager.SaveDisabledModulePaths(disabled_modules);
-      [[fallthrough]];
+      app_->DisableDownloadForModule(module->file_path());
+      break;
     }
     case Result::kStop:
       break;
